refactor assets to files
